- All uses of variables bound in a single scope (except formal parameters) are translated into references to their location in the appropriate environment instead of lookups. Use the get_location methods in VarInfo and VarBinding. (Function uses referring to a single definition go a step further; see below.)
- Compiling call sites: when the call graph indicates that only one
function definition is possible, RCC now generates a direct reference
to the closure, bypassing variables altogether.

- VarBinding objects can legitimately contain an empty list of scopes.
These represent names with no binding in user code, which used to be
lumped in with names bound at the global scope. This can occur when a
name is bound only in the R environment, as library/builtin/special
names, or when the variable is just plain unbound.
VarBinding::is_unbound() returns true in these cases.

- added FuncInfo::is_arg(SEXP)

- added is_library in Utils

- fixed visibility bug in op_return


git-svn-id: file:///home/john/Code/repo-dumps/rcc/trunk@200 925864f1-7c04-44ca-bbaa-d8d9a55e96f5