* RCC now compiles most programs that redefine names in the internal R environment. Names representing builtin objects of BUILTINSXP type (e.g. arithmetic operators) and library objects of CLOSXP type (e.g. qnorm) can be redefined; redefinition of SPECIALSXPs (e.g. "while") is thankfully still prohibited and will trigger a trivial compile.
* Reworked RCC's idea of the R scope tree: names in the internal R
environment are bound in a separate scope "behind" the scope tree,
instead of the same global scope accessible by user programs.

* Accordingly, created LexicalScope class to represent a lexical
scope, which may or may not be associated with a function definition.
Most of the changes are to SymbolTable, op_lang, and VarBinding.

* Improved some dump routines.

* Fixed a bug where expressions in parentheses were mishandled in call
graph building.

* Fixed a bug with double-arrow (<<-) assignments to internal names.

* Made op_var_def fail gracefully if it fails to get a location.

* Changed the "well-behaved" assertion so that redefinitions of
builtin and library names are no longer prohibited.

* The "redef" regression tests are now well-behaved. Added some new
regression tests.

Note: one regression having to do with anonymous functions in the call
graph. I'm working on it.


git-svn-id: file:///home/john/Code/repo-dumps/rcc/trunk@205 925864f1-7c04-44ca-bbaa-d8d9a55e96f5